feat: Move UI components to kmpuiviews and Update Settings
This commit moves several UI components, such as `InfiniteListHandler`, `SelectableMiniPalette`, `DiscordIcon`, and `GitHubIcon`, from the `UIViews` module to the `kmpuiviews` module, for better code organization and potential cross-platform use. It also updates settings fragments and onboarding content to use the new location of these components.

-   **UI Components Migration:**
    -   Moved `InfiniteListHandler` to `kmpuiviews` and removed it from `UIViews`.
    - Moved `SelectableMiniPalette` and `ThemeItem` to `kmpuiviews` and removed them from `UIViews`.
    - Moved `DiscordIcon` and `GithubIcon` to `kmpuiviews` and removed them from `UIViews`.
    - Added a `seedColor` extension property to `ThemeColor` in `kmpuiviews`.
-   **Settings and Onboarding Updates:**
    -   Updated `RecentFragment` and `AllFragment` to use the new `InfiniteListHandler` from `kmpuiviews`.
    -   Updated `GeneralSettings` to use the new `ThemeItem` and `seedColor` from `kmpuiviews`.
    -   Updated `InfoSettings` to use the new `Discord` and `Github` icons from `kmpuiviews`.
-  - Updated `OnboardingContent` to use the new `ThemeItem` and `seedColor` from `kmpuiviews`.
+    - Updated `OnboardingContent` to use the new `ThemeItem` and `seedColor` from `kmpuiviews`.
